Moving some inline helpers to source files
...for faster compilation, with negligible performance hit.
