+++ Framework/CMakeLists.txt
@@ -55,6 +55,8 @@ add_library(O2QualityControl
   src/TaskInterface.cxx
   src/UserCodeInterface.cxx
   src/RepositoryBenchmark.cxx
+  src/RepoPathUtils.cxx
+  src/stringUtils.cxx
   src/InfrastructureGenerator.cxx
   src/InfrastructureSpecReader.cxx
   src/Check.cxx
