+++ Framework/include/QualityControl/runnerUtils.h
@@ -18,21 +18,11 @@
 #define QUALITYCONTROL_RUNNERUTILS_H
 
 #include <string>
-#include <Configuration/ConfigurationFactory.h>
-#include <Common/Exceptions.h>
+#include <Framework/ServiceRegistryRef.h>
 #include <Framework/RawDeviceService.h>
-#include <Framework/DeviceSpec.h>
 #include <fairmq/Device.h>
-#include <Framework/ConfigParamRegistry.h>
-#include <QualityControl/QcInfoLogger.h>
-#include <boost/property_tree/json_parser.hpp>
-#include <CommonUtils/StringUtils.h>
+#include "QualityControl/QcInfoLogger.h"
 #include "QualityControl/Activity.h"
-#include <regex>
-#include "QualityControl/Bookkeeping.h"
-
-// TODO: most of these could be moved to a source file with minimal performance impact
-//  and noticable compilation time improvement (we bring boost, regex, FairMQ headers to 10 other source files)
 
 namespace o2::quality_control::core
 {
@@ -44,38 +34,12 @@ namespace o2::quality_control::core
  * @param config
  * @return The name of the first task in the config file.
  */
-inline std::string getFirstTaskName(std::string configurationSource)
-{
-  auto config = o2::configuration::ConfigurationFactory::getConfiguration(configurationSource);
-
-  for (const auto& task : config->getRecursive("qc.tasks")) {
-    return task.first; // task name;
-  }
-
-  throw;
-}
-
-inline std::string getFirstCheckName(std::string configurationSource)
-{
-  auto config = o2::configuration::ConfigurationFactory::getConfiguration(configurationSource);
-
-  if (config->getRecursive("qc").count("checks")) {
-    for (const auto& check : config->getRecursive("qc.checks")) {
-      return check.first; // check name;
-    }
-  }
-
-  BOOST_THROW_EXCEPTION(AliceO2::Common::ObjectNotFoundError() << AliceO2::Common::errinfo_details("No checks defined"));
-}
-
-inline bool hasChecks(std::string configSource)
-{
-  auto config = o2::configuration::ConfigurationFactory::getConfiguration(configSource);
-  return config->getRecursive("qc").count("checks") > 0;
-}
+std::string getFirstTaskName(const std::string& configurationSource);
+std::string getFirstCheckName(const std::string& configurationSource);
+bool hasChecks(const std::string& configSource);
 
 template <typename T> // TODO we should probably limit T to numbers somehow
-inline T computeActivityField(framework::ServiceRegistryRef services, const std::string& name, T fallbackNumber = 0)
+T computeActivityField(framework::ServiceRegistryRef services, const std::string& name, T fallbackNumber = 0)
 {
   int result = 0;
 
@@ -95,66 +59,13 @@ inline T computeActivityField(framework::ServiceRegistryRef services, const std:
   return result;
 }
 
-inline Activity computeActivity(framework::ServiceRegistryRef services, const Activity& fallbackActivity)
-{
-  auto runNumber = computeActivityField<int>(services, "runNumber", fallbackActivity.mId);
-  auto runType = computeActivityField<int>(services, "runType", fallbackActivity.mType);
-  auto run_start_time_ms = computeActivityField<unsigned long>(services, "runStartTimeMs", fallbackActivity.mValidity.getMin());
-  auto run_stop_time_ms = computeActivityField<unsigned long>(services, "runEndTimeMs", fallbackActivity.mValidity.getMax());
-  auto partitionName = services.get<framework::RawDeviceService>().device()->fConfig->GetProperty<std::string>("environment_id", fallbackActivity.mPartitionName);
-  auto periodName = services.get<framework::RawDeviceService>().device()->fConfig->GetProperty<std::string>("lhcPeriod", fallbackActivity.mPeriodName);
-  auto fillNumber = computeActivityField<int>(services, "fillInfoFillNumber", fallbackActivity.mFillNumber);
-  auto beam_type = services.get<framework::RawDeviceService>().device()->fConfig->GetProperty<std::string>("fillInfoBeamType", fallbackActivity.mBeamType);
+Activity computeActivity(framework::ServiceRegistryRef services, const Activity& fallbackActivity);
 
-  Activity activity(
-    runNumber,
-    runType,
-    periodName,
-    fallbackActivity.mPassName,
-    fallbackActivity.mProvenance,
-    { run_start_time_ms, run_stop_time_ms },
-    beam_type,
-    partitionName,
-    fillNumber);
+std::string indentTree(int level);
+void printTree(const boost::property_tree::ptree& pt, int level = 0);
 
-  return activity;
-}
-
-inline std::string indentTree(int level)
-{
-  return std::string(level * 2, ' ');
-}
-
-inline void printTree(const boost::property_tree::ptree& pt, int level = 0)
-{
-  std::stringstream ss;
-  boost::property_tree::json_parser::write_json(ss, pt);
-  for (std::string line; std::getline(ss, line, '\n');)
-    ILOG(Debug, Trace) << line << ENDM;
-}
-
-inline std::vector<std::pair<std::string, std::string>> parseOverrideValues(const std::string& input)
-{
-  std::vector<std::pair<std::string, std::string>> keyValuePairs;
-  for (const auto& keyValueToken : utils::Str::tokenize(input, ';', true)) {
-    auto keyValue = utils::Str::tokenize(keyValueToken, '=', true);
-    if (keyValue.size() == 1) {
-      keyValuePairs.emplace_back(keyValue[0], "");
-    } else if (keyValue.size() == 2) {
-      keyValuePairs.emplace_back(keyValue[0], keyValue[1]);
-    } else {
-      throw std::runtime_error("Token '" + keyValueToken + "' in the --override-values argument is malformed, use key=value.");
-    }
-  }
-  return keyValuePairs;
-}
-
-inline void overrideValues(boost::property_tree::ptree& tree, std::vector<std::pair<std::string, std::string>> keyValues)
-{
-  for (const auto& [key, value] : keyValues) {
-    tree.put(key, value);
-  }
-}
+std::vector<std::pair<std::string, std::string>> parseOverrideValues(const std::string& input);
+void overrideValues(boost::property_tree::ptree& tree, const std::vector<std::pair<std::string, std::string>>& keyValues);
 
 /**
  * template the param infologgerDiscardFile (_ID_->[device-id])
@@ -162,16 +73,7 @@ inline void overrideValues(boost::property_tree::ptree& tree, std::vector<std::p
  * @param iCtx
  * @return
  */
-inline std::string templateILDiscardFile(std::string& originalFile, framework::InitContext& iCtx)
-{
-  try {
-    auto& deviceSpec = iCtx.services().get<o2::framework::DeviceSpec const>();
-    return std::regex_replace(originalFile, std::regex("_ID_"), deviceSpec.id);
-  } catch (...) {
-    ILOG(Error, Devel) << "exception caught and swallowed in templateILDiscardFile : " << boost::current_exception_diagnostic_information() << ENDM;
-  }
-  return originalFile;
-}
+std::string templateILDiscardFile(std::string& originalFile, framework::InitContext& iCtx);
 
 uint64_t getCurrentTimestamp();
